+++ src/main/java/gil/server/ParametersController.java
@@ -3,18 +3,23 @@ package gil.server;
 import java.io.UnsupportedEncodingException;
 import java.util.function.BiFunction;
 
-public class RouteNotFoundController {
+public class ParametersController {
     public static BiFunction<Request, Response, Response> get =
             (request, response) -> {
+                String responseParameters = request.getParameters().toString();
+                int lengthOfStringWithNoParameters = 2;
+                Boolean requestHasParameters = responseParameters.length() > lengthOfStringWithNoParameters;
+                String body = requestHasParameters ? responseParameters : "";
                 response.setProtocol("HTTP/1.1");
-                response.setStatusCode("404");
-                response.setReasonPhrase("Not Found");
+                response.setStatusCode("200");
+                response.setReasonPhrase("OK");
                 response.setContentType("text/html; charset=UTF-8");
                 try {
-                    response.setBody("");
+                    response.setBody(body);
                 } catch (UnsupportedEncodingException e) {
                     e.printStackTrace();
                 }
+
                 return response;
             };
 }
